+++ crates/cli/src/session/format/binary.rs
@@ -10,12 +10,12 @@ use super::MessageFormatter;
 pub struct MsgBinaryFormatter {}
 
 impl MessageFormatter for MsgBinaryFormatter {
-    fn write_msg<M>(&mut self, mut writer: impl std::io::Write, msg: &M) -> anyhow::Result<()>
+    fn write_msg<M>(&mut self, mut writer: impl std::io::Write, msg: &M) -> anyhow::Result<bool>
     where
         M: LogMessage,
     {
         msg.to_writer(&mut writer)
             .context("Error while writing binary message")?;
-        Ok(())
+        Ok(true)
     }
 }